+++ dst/app/routes/places/screeningRoom.js
@@ -15,8 +15,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const chevre = require("@chevre/domain");
 const createDebug = require("debug");
 const express_1 = require("express");
-// tslint:disable-next-line:no-submodule-imports
-const check_1 = require("express-validator/check");
+const express_validator_1 = require("express-validator");
 const http_status_1 = require("http-status");
 const mongoose = require("mongoose");
 const authentication_1 = require("../../middlewares/authentication");
@@ -29,27 +28,27 @@ screeningRoomRouter.use(authentication_1.default);
  * 作成
  */
 screeningRoomRouter.post('', permitScopes_1.default(['admin']), ...[
-    check_1.body('project')
+    express_validator_1.body('project')
         .not()
         .isEmpty()
         .withMessage(() => 'Required'),
-    check_1.body('branchCode')
+    express_validator_1.body('branchCode')
         .not()
         .isEmpty()
         .withMessage(() => 'Required'),
-    check_1.body('name')
+    express_validator_1.body('name')
         .not()
         .isEmpty()
         .withMessage(() => 'Required'),
-    check_1.body('containedInPlace.branchCode')
+    express_validator_1.body('containedInPlace.branchCode')
         .not()
         .isEmpty()
         .withMessage(() => 'Required')
         .isString(),
-    check_1.body('additionalProperty')
+    express_validator_1.body('additionalProperty')
         .optional()
         .isArray(),
-    check_1.body('openSeatingAllowed')
+    express_validator_1.body('openSeatingAllowed')
         .optional()
         .isBoolean()
         .toBoolean()
@@ -228,28 +227,29 @@ screeningRoomRouter.get('', permitScopes_1.default(['admin']), validator_1.defau
 /**
  * 更新
  */
+// tslint:disable-next-line:use-default-type-parameter
 screeningRoomRouter.put('/:branchCode', permitScopes_1.default(['admin']), ...[
-    check_1.body('project')
+    express_validator_1.body('project')
         .not()
         .isEmpty()
         .withMessage(() => 'Required'),
-    check_1.body('branchCode')
+    express_validator_1.body('branchCode')
         .not()
         .isEmpty()
         .withMessage(() => 'Required'),
-    check_1.body('name')
+    express_validator_1.body('name')
         .not()
         .isEmpty()
         .withMessage(() => 'Required'),
-    check_1.body('containedInPlace.branchCode')
+    express_validator_1.body('containedInPlace.branchCode')
         .not()
         .isEmpty()
         .withMessage(() => 'Required')
         .isString(),
-    check_1.body('additionalProperty')
+    express_validator_1.body('additionalProperty')
         .optional()
         .isArray(),
-    check_1.body('openSeatingAllowed')
+    express_validator_1.body('openSeatingAllowed')
         .optional()
         .isBoolean()
         .toBoolean()
@@ -295,12 +295,13 @@ screeningRoomRouter.put('/:branchCode', permitScopes_1.default(['admin']), ...[
 /**
  * 削除
  */
+// tslint:disable-next-line:use-default-type-parameter
 screeningRoomRouter.delete('/:branchCode', permitScopes_1.default(['admin']), ...[
-    check_1.body('project')
+    express_validator_1.body('project')
         .not()
         .isEmpty()
         .withMessage(() => 'Required'),
-    check_1.body('containedInPlace.branchCode')
+    express_validator_1.body('containedInPlace.branchCode')
         .not()
         .isEmpty()
         .withMessage(() => 'Required')