+++ src/app/routes/places/screeningRoomSection.ts
@@ -3,8 +3,9 @@
  */
 import * as chevre from '@chevre/domain';
 import { Router } from 'express';
-// tslint:disable-next-line:no-submodule-imports
-import { body } from 'express-validator/check';
+// tslint:disable-next-line:no-implicit-dependencies
+import { ParamsDictionary } from 'express-serve-static-core';
+import { body } from 'express-validator';
 import { CREATED, NO_CONTENT } from 'http-status';
 import * as mongoose from 'mongoose';
 
@@ -264,7 +265,8 @@ screeningRoomSectionRouter.get(
 /**
  * 更新
  */
-screeningRoomSectionRouter.put(
+// tslint:disable-next-line:use-default-type-parameter
+screeningRoomSectionRouter.put<ParamsDictionary>(
     '/:branchCode',
     permitScopes(['admin']),
     ...[
@@ -372,7 +374,8 @@ screeningRoomSectionRouter.put(
 /**
  * 削除
  */
-screeningRoomSectionRouter.delete(
+// tslint:disable-next-line:use-default-type-parameter
+screeningRoomSectionRouter.delete<ParamsDictionary>(
     '/:branchCode',
     permitScopes(['admin']),
     ...[